# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/create_cross_sec_area": {
            "post": {
                "description": "Создает новую площадь сечения",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "welding_info"
                ],
                "summary": "Вставить новую площадь сечения",
                "parameters": [
                    {
                        "description": "Значение площади сечения",
                        "name": "area_value",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.InsertCrossSecAreaRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Площадь сечения успешно создана",
                        "schema": {
                            "$ref": "#/definitions/handlers.CrossSecAreaResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный формат данных или отсутствуют обязательные поля",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/create_katet": {
            "post": {
                "description": "Добавляет в базу данных новый катет(толщину)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "welding_info"
                ],
                "summary": "Вставить новый катет(толщину)",
                "parameters": [
                    {
                        "description": "Значение катета",
                        "name": "katet_value",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.InsertKatetRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Катет успешно создан",
                        "schema": {
                            "$ref": "#/definitions/handlers.SeamTypeResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный формат данных или отсутствуют обязательные поля",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/create_seam_type": {
            "post": {
                "description": "Создает новый тип шва с указанным именем.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "welding_info"
                ],
                "summary": "Вставить новый тип шва",
                "parameters": [
                    {
                        "description": "Название типа шва",
                        "name": "seam_type_name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.InsertSeamTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Тип шва успешно создан",
                        "schema": {
                            "$ref": "#/definitions/handlers.SeamTypeResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный формат данных или отсутствуют обязательные поля",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cross_sec_area": {
            "get": {
                "description": "Возвращает информацию о площади сечения шва на основе материала, типа сварки, катета(толщины), типа шва",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "welding_info"
                ],
                "summary": "Получение информации о площади сечения шва",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название материала",
                        "name": "material_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название типа сварки",
                        "name": "welding_type_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Значение катета(толщины)",
                        "name": "katet_value",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название типа шва",
                        "name": "seam_type_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CrossSecArea"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/katets": {
            "get": {
                "description": "Возвращает информацию о типах сварки для заданного материала и типа сварки",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "welding_info"
                ],
                "summary": "Получение информации о катетах(толщине) металла",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название материала",
                        "name": "material_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название типа сварки",
                        "name": "welding_type_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Katets"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/seam_type": {
            "get": {
                "description": "Возвращает информацию о типе шва по заданным материалу, типу сварки и катету(толщине)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "welding_info"
                ],
                "summary": "Получение информации о типе шва",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название материала",
                        "name": "material_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название типа сварки",
                        "name": "welding_type_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Катет(толщина) материала",
                        "name": "katet_value",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.SeamTypes"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.CrossSecAreaResponse": {
            "type": "object",
            "properties": {
                "area_value": {
                    "type": "string"
                }
            }
        },
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "handlers.InsertCrossSecAreaRequest": {
            "type": "object",
            "properties": {
                "area_value": {
                    "type": "number"
                }
            }
        },
        "handlers.InsertKatetRequest": {
            "type": "object",
            "required": [
                "katet_value"
            ],
            "properties": {
                "katet_value": {
                    "type": "string"
                }
            }
        },
        "handlers.InsertSeamTypeRequest": {
            "type": "object",
            "required": [
                "seam_type_name"
            ],
            "properties": {
                "seam_type_name": {
                    "type": "string"
                }
            }
        },
        "handlers.SeamTypeResponse": {
            "type": "object",
            "properties": {
                "seam_type_name": {
                    "type": "string"
                }
            }
        },
...
        "models.CrossSecArea": {
            "type": "object",
            "properties": {
                "area_value": {
                    "type": "number"
                }
            }
        },
...
        "models.Katets": {
            "type": "object",
            "properties": {
                "katet_value": {
                    "type": "string"
                }
            }
        },
...
        "models.SeamTypes": {
            "type": "object",
            "properties": {
                "seam_type_name": {
                    "type": "string"
                }
            }
        },
...
    }
}